Add Pro controllers colors change
Unfortunatelly, no hw no test. So I took users testing for granted..

Had to change my Joy-Con to be recognisable as a Pro Controller to test that the buttons color work on Pro controller on Switch and that the default white is actually to distinguish them from the black body.

Note: Now Pro buttons WHITE color will be previewed as Grey.