[IR/NFC] Fix IR/NFC in 3.89 jc fw
Still it's unknown what the values x04/x05 and x12/x18 are.

Thoughts for food:
Maybe it's the MCU version? 4.12 in <=3.86 and 5.18 in 3.89?

So the switch requested 3.09 mcu version is the minimum?

The joy-con got an mcu fw update also? How much time the update took? The PatchRam fw update in SPI is very fast.

Or the PatchRam is able to live patch the stm32 MCU firmware also?